nicework/requirements.txt
asgiref==3.4.1
Django==3.2.13
lunardate==0.2.0
pytimekr==0.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
wheel==0.37.1
python-dateutil==2.8.2
six==1.16.0